requirements.txt
fastmcp>=2.0.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pydantic>=2.0.0
numpy>=1.26.0
pytest>=8.0.0